sandbox: Split the StandaloneApiServer from the StandaloneIndexerServer.
