+++ ledger/api-server-damlonx/reference-v2/src/main/scala/com/daml/ledger/api/server/damlonx/reference/v2/cli/Cli.scala
@@ -6,7 +6,7 @@ import java.io.File
 
 import com.digitalasset.daml.lf.data.Ref
 import com.digitalasset.ledger.api.tls.TlsConfiguration
-import com.digitalasset.platform.index.config.Config
+import com.digitalasset.platform.apiserver.Config
 import scopt.Read
 
 object Cli {
