+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/apiserver/ApiServices.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2019 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.ledger.server.apiserver
+package com.digitalasset.platform.apiserver
 
 import akka.stream.ActorMaterializer
 import com.codahale.metrics.MetricRegistry
