+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/apiserver/JdbcIndex.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2019 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.platform.index
+package com.digitalasset.platform.apiserver
 
 import akka.NotUsed
 import akka.stream.Materializer
