+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/indexer/RecoveringIndexer.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2019 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.platform.index
+package com.digitalasset.platform.indexer
 
 import java.util.concurrent.atomic.{AtomicBoolean, AtomicReference}
